Bump to arrow/parquet 55 rc and remove object_store integration (#718)
* Bump to arrow/parquet 55 and remove object_store

* Bump to parquet 55

* Remove object_store usage from parquet

* Use upstream coalesce_ranges for simplicity

* use internal coalesce byte size

* unsafe impl sync

+++ src/common/fetch.rs
@@ -33,7 +33,7 @@ pub fn range_from_start(start: u64) -> String {
     format!("bytes={}-", start)
 }
 
-pub fn range_from_end(length: u64) -> String {
+pub fn range_from_end(length: usize) -> String {
     format!("bytes=-{}", length)
 }
